Module zipp::consts

Some const-variable flag const variable flags

Constants

ZIP_AFL_RDONLY

ZIP_CHECKCONS

OpenFlag Perform additional stricter consistency checks on the archive, and error if they fail.

ZIP_CM_BZIP2

Compression compressed using BZIP2 algorithm

ZIP_CM_DEFAULT

Compression better of deflate or store

ZIP_CM_DEFLATE

Compression deflated

ZIP_CM_DEFLATE64

Compression deflate64

ZIP_CM_IMPLODE

Compression imploded

ZIP_CM_JPEG

Compression Compressed Jpeg data

ZIP_CM_LZ77

Compression IBM LZ77 z Architecture

ZIP_CM_LZMA

Compression LZMA (EFS)

ZIP_CM_LZMA2

Compression

ZIP_CM_PKWARE_IMPLODE

Compression PKWARE imploding

ZIP_CM_PPMD

Compression PPMd version I, Rev 1

ZIP_CM_REDUCE_1

Compression reduced with factor 1

ZIP_CM_REDUCE_2

Compression reduced with factor 2

ZIP_CM_REDUCE_3

Compression reduced with factor 3

ZIP_CM_REDUCE_4

Compression reduced with factor 4

ZIP_CM_SHRINK

Compression shrunk

ZIP_CM_STORE

Compression stored (uncompressed)

ZIP_CM_TERSE

Compression compressed using IBM TERSE

ZIP_CM_WAVPACK

Compression WavPack compressed data

ZIP_CM_XZ

Compression XZ compressed data

ZIP_CM_ZSTD

Compression Zstandard compressed data

ZIP_CREATE

OpenFlag Create the archive if it does not exist.

ZIP_EM_AES_128

Encryption AES-128

ZIP_EM_AES_192

Encryption AES-192

ZIP_EM_AES_256

Encryption AES-256

ZIP_EM_NONE

Encryption not encrypted

ZIP_EM_TRAD_PKWARE

Encryption traditional PKWARE encryption

ZIP_EM_UNKNOWN

Encryption unknown algorithm

ZIP_ER_CANCELLED

Error Operation cancelled.

ZIP_ER_CHANGED

Error Entry has been changed.

ZIP_ER_CLOSE

Error Closing zip archive failed.

ZIP_ER_COMPNOTSUPP

Error Compression method not supported.

ZIP_ER_COMPRESSED_DATA

Error Compressed data invalid.

ZIP_ER_CRC

Error CRC error.

ZIP_ER_DELETED

Error Entry has been deleted.

ZIP_ER_ENCRNOTSUPP

Error Encryption method not supported.

ZIP_ER_EOF

Error Premature end of file.

ZIP_ER_EXISTS

Error File already exists.

ZIP_ER_INCONS

Error Zip archive inconsistent.

ZIP_ER_INTERNAL

Error Internal error.

ZIP_ER_INUSE

Error Resource still in use.

ZIP_ER_INVAL

Error Invalid argument.

ZIP_ER_MEMORY

Error Malloc failure.

ZIP_ER_MULTIDISK

Error Multi-disk zip archives not supported.

ZIP_ER_NOENT

Error No such file.

ZIP_ER_NOPASSWD

Error No password provided.

ZIP_ER_NOZIP

Error Not a zip archive.

ZIP_ER_OK

Error No error.

ZIP_ER_OPEN

Error Can’t open file.

ZIP_ER_OPNOTSUPP

Error Operation not supported.

ZIP_ER_RDONLY

Error Read-only archive.

ZIP_ER_READ

Error Read error.

ZIP_ER_REMOVE

Error Can’t remove file.

ZIP_ER_RENAME

Error Renaming temporary file failed.

ZIP_ER_SEEK

Error Seek error.

ZIP_ER_TELL

Error Tell error.

ZIP_ER_TMPOPEN

Error Failure to create temporary file.

ZIP_ER_WRITE

Error Write error.

ZIP_ER_WRONGPASSWD

Error Wrong password provided.

ZIP_ER_ZIPCLOSED

Error Containing zip archive was closed.

ZIP_ER_ZLIB

Error Zlib error.

ZIP_ET_LIBZIP

sys_err is libzip error code

ZIP_ET_NONE

sys_err unused

ZIP_ET_SYS

sys_err is errno

ZIP_ET_ZLIB

sys_err is zlib error code

ZIP_EXCL

OpenFlag Error if archive already exists.

ZIP_EXTRA_FIELD_ALL

ZIP_EXTRA_FIELD_NEW

ZIP_FL_CENTRAL

in central directory

ZIP_FL_COMPRESSED

read compressed data

ZIP_FL_ENCRYPTED

read encrypted data (implies ZIP_FL_COMPRESSED)

ZIP_FL_ENC_CP437

string is CP437 encoded

ZIP_FL_ENC_GUESS

guess string encoding (is default)

ZIP_FL_ENC_RAW

get unmodified string

ZIP_FL_ENC_STRICT

follow specification strictly

ZIP_FL_ENC_UTF_8

string is UTF-8 encoded

ZIP_FL_LOCAL

in local header

ZIP_FL_NOCASE

ignore case on name lookup

ZIP_FL_NODIR

ignore directory component

ZIP_FL_OVERWRITE

if file with name exists, overwrite (replace) it

ZIP_FL_RECOMPRESS

force recompression of data

ZIP_FL_UNCHANGED

use original data, ignoring changes

ZIP_RDONLY

OpenFlag Open archive in read-only mode.

ZIP_STAT_ALL

Stat get stat all field

ZIP_STAT_COMP_METHOD

Stat get stat comp_method

ZIP_STAT_COMP_SIZE

Stat get stat comp_size

ZIP_STAT_CRC

Stat get stat crc

ZIP_STAT_ENCRYPTION_METHOD

Stat get stat encryption_method

ZIP_STAT_FLAGS

Stat get stat flags

ZIP_STAT_INDEX

Stat get stat index

ZIP_STAT_MTIME

Stat get stat mtime

ZIP_STAT_NAME

Stat get stat name

ZIP_STAT_SIZE

Stat get stat size

ZIP_TRUNCATE

OpenFlag If archive exists, ignore its current contents. In other words, handle it the same way as an empty archive.